feat: Add Navbar component for website navigation
This commit adds a new component called "Navbar.svelte" to the project. The Navbar component is responsible for displaying the navigation menu on the website. It includes a list of menu items such as "Acerca", "Ayuda", and "Empieza un Proyecto". The component also imports and uses the ConnectPlug component for wallet connection. Some commented out code and styles have been removed from the Navbar.svelte file.