feat: Add carbon-components-svelte for layout styling
This commit adds the `carbon-components-svelte` library to the project, specifically for the `+layout.svelte` file. The library provides pre-styled components and CSS classes that can be used to enhance the layout and styling of the application. This addition will improve the visual appearance and user experience of the application's layout.